[Store] Split DummyClient and offload unit tests
