+++ mooncake-store/tests/dummy_client/main.cpp
@@ -1,0 +1,8 @@
+#include <gflags/gflags.h>
+#include <gtest/gtest.h>
+
+int main(int argc, char** argv) {
+    ::testing::InitGoogleTest(&argc, argv);
+    gflags::ParseCommandLineFlags(&argc, &argv, false);
+    return RUN_ALL_TESTS();
+}
