+++ mooncake-store/tests/master_service/offload/fixture.h
@@ -1,0 +1,130 @@
+#pragma once
+
+// Shared fixture and helpers for MasterService offload tests.
+
+#include "master_service.h"
+
+#include <glog/logging.h>
+#include <gtest/gtest.h>
+
+#include <atomic>
+#include <chrono>
+#include <memory>
+#include <string>
+#include <thread>
+#include <vector>
+
+#include "types.h"
+
+namespace mooncake::test {
+
+class OffloadOnEvictTest : public ::testing::Test {
+   protected:
+    void SetUp() override {
+        google::InitGoogleLogging("OffloadOnEvictTest");
+        FLAGS_logtostderr = true;
+    }
+
+    void TearDown() override { google::ShutdownGoogleLogging(); }
+
+    static constexpr size_t kDefaultSegmentBase = 0x300000000;
+
+    Segment MakeSegment(std::string name, size_t base, size_t size) const {
+        Segment segment;
+        segment.id = generate_uuid();
+        segment.name = std::move(name);
+        segment.base = base;
+        segment.size = size;
+        segment.te_endpoint = segment.name;
+        return segment;
+    }
+
+    struct MountedSegmentContext {
+        UUID segment_id;
+        UUID client_id;
+    };
+
+    MountedSegmentContext PrepareSegment(MasterService& service,
+                                         std::string name, size_t base,
+                                         size_t size) const {
+        Segment segment = MakeSegment(std::move(name), base, size);
+        UUID client_id = generate_uuid();
+        auto mount_result = service.MountSegment(segment, client_id);
+        EXPECT_TRUE(mount_result.has_value());
+        return {.segment_id = segment.id, .client_id = client_id};
+    }
+
+    // Put an object and complete it.
+    void PutObject(MasterService& service, const UUID& client_id,
+                   const std::string& key, size_t size = 1024) {
+        ReplicateConfig config;
+        config.replica_num = 1;
+        auto put_start =
+            service.PutStart(client_id, key, "default", size, config);
+        ASSERT_TRUE(put_start.has_value()) << "PutStart failed for key=" << key;
+        auto put_end =
+            service.PutEnd(client_id, key, "default", ReplicaType::MEMORY);
+        ASSERT_TRUE(put_end.has_value()) << "PutEnd failed for key=" << key;
+    }
+
+    // Drain the offload queue via OffloadObjectHeartbeat.
+    std::unordered_map<std::string, int64_t> DrainOffloadQueue(
+        MasterService& service, const UUID& client_id) {
+        auto res = service.OffloadObjectHeartbeat(client_id, true);
+        if (!res) {
+            return {};
+        }
+        std::unordered_map<std::string, int64_t> queued;
+        for (const auto& task : res.value()) {
+            queued[task.key] = task.size;
+        }
+        return queued;
+    }
+
+    template <typename Predicate>
+    void WaitUntil(
+        Predicate&& predicate,
+        std::chrono::milliseconds timeout = std::chrono::milliseconds(4000),
+        std::chrono::milliseconds interval =
+            std::chrono::milliseconds(50)) const {
+        const auto deadline = std::chrono::steady_clock::now() + timeout;
+        while (std::chrono::steady_clock::now() < deadline) {
+            if (predicate()) {
+                return;
+            }
+            std::this_thread::sleep_for(interval);
+        }
+        EXPECT_TRUE(predicate());
+    }
+
+    // Fill a segment until PutStart fails, triggering eviction.
+    // Returns the number of successful puts.
+    int FillSegmentUntilEviction(MasterService& service, const UUID& client_id,
+                                 const std::string& key_prefix,
+                                 size_t object_size, int max_puts) {
+        int success_puts = 0;
+        for (int i = 0; i < max_puts; ++i) {
+            std::string key = key_prefix + std::to_string(i);
+            ReplicateConfig config;
+            config.replica_num = 1;
+            auto result = service.PutStart(client_id, key, "default",
+                                           object_size, config);
+            if (result.has_value()) {
+                auto end = service.PutEnd(client_id, key, "default",
+                                          ReplicaType::MEMORY);
+                EXPECT_TRUE(end.has_value());
+                success_puts++;
+            } else {
+                // Wait for eviction to process
+                std::this_thread::sleep_for(std::chrono::milliseconds(50));
+            }
+        }
+        return success_puts;
+    }
+};
+
+// =============================================================================
+// Combo A: Default config (offload at PutEnd)
+// =============================================================================
+
+}  // namespace mooncake::test
